files not added bc of bug in Github Desktop
some files were too big and stopped upload of all later files. The bug with Github Desktop/git on command line is that it show that there are no changes, but these files are missing in the master repository and yet are found in my local copy. Uploading these manually on the web browser